refactor : 스터디 검색 기능 병합
